Add support for ADS1115
Add transparent support for ADS1115 by measuring the conversion time.

The ADS1115 is a slower, 16-bit version of the 12-bit ADS1015.

Since ADS1115 is roughly 16x slower than ADS1105 with the same settings, we can detect which chip our measurements are coming from an compensate accordingly.

Thanks to Niko for this - somewhat jokey - stroke of genius.